Remove useless float conversions and cleanup

+++ TimelineView/src/main/java/com/lriccardo/timelineview/TimelineDecorator.kt
@@ -8,12 +8,12 @@ import androidx.recyclerview.widget.RecyclerView
 
 class TimelineDecorator(
     val indicatorRadius: Float = 24f,
-    val lineWidth: Float = indicatorRadius/1.61f,
-    val padding: Float = indicatorRadius*2,
+    val lineWidth: Float = indicatorRadius / 1.61f,
+    val padding: Float = indicatorRadius * 2,
     val position: Position = Position.Left
 ) : RecyclerView.ItemDecoration() {
 
-    val width = ((indicatorRadius * 2)+(padding*2))
+    val width = ((indicatorRadius * 2) + (padding * 2))
 
     enum class Position {
         Left,
@@ -26,7 +26,7 @@ class TimelineDecorator(
         parent: RecyclerView,
         s: RecyclerView.State
     ) {
-        when(position){
+        when (position) {
             Position.Left ->
                 rect.left = width.toInt()
             Position.Right ->
@@ -43,21 +43,21 @@ class TimelineDecorator(
 
             val timelineView = TimelineView(context = parent.context)
             timelineView.setType(itemPosition, parent.adapter?.itemCount ?: -1)
-            timelineView.indicatorRadius = indicatorRadius.toFloat()
-            timelineView.lineWidth = lineWidth.toFloat()
+            timelineView.indicatorRadius = indicatorRadius
+            timelineView.lineWidth = lineWidth
 
             timelineView.measure(
                 View.MeasureSpec.getSize(width.toInt()),
                 View.MeasureSpec.getSize(it.measuredHeight)
             )
             c.save()
-            when(position){
+            when (position) {
                 Position.Left -> {
-                    c.translate(padding.toFloat(), it.top.toFloat())
+                    c.translate(padding, it.top.toFloat())
                     timelineView.layout(0, 0, timelineView.measuredWidth, it.measuredHeight)
                 }
                 Position.Right -> {
-                    c.translate(it.measuredWidth+padding.toFloat(), it.top.toFloat())
+                    c.translate(it.measuredWidth + padding, it.top.toFloat())
                     timelineView.layout(0, 0, timelineView.measuredWidth, it.measuredHeight)
                 }
             }
